+++ lib/src/enums/popup_trigger_behavior.dart
@@ -1,0 +1,8 @@
+/// [PopupTriggerBehavior] is used to specify the trigger behavior of the info popup.
+enum PopupTriggerBehavior {
+  /// [onTap] is used to trigger the popup on tap.
+  onTap,
+
+  /// [manuel] is used to trigger the popup manually with controller.
+  manuel
+}